Add files via upload
KEEPING THE ENVIRONMENT CLEAN

Idea of this game is inculcate a habit in the user about contributing in maintaining a clean environment and thereby, giving a beautiful world to future generations. If my game can ignite a thought in user’s mind about keeping environment clean, it would be a great achievement for me.